+++ assigner/commands/status.py
@@ -35,7 +35,7 @@ def status(conf, args):
 
     output = PrettyTable([
         "#", "Sec", "SID", "Name", "Status", "Branches",
-        "HEAD", "Last Commit Author", "Last Commit Time"])
+        "HEAD", "Last Commit Author", "Last Push Time"])
     output.align["Name"] = "l"
     output.align["Last Commit Author"] = "l"
 
@@ -85,10 +85,9 @@ def status(conf, args):
             if branches:
                 row[5] = ", ".join([b["name"] for b in branches])
 
-            commits = repo.list_commits()
+            head = repo.get_last_HEAD_commit()
 
-            if commits:
-                head = commits[0]
+            if head:
                 row[6] = head["short_id"]
                 row[7] = head["author_name"]
                 created_at = head["created_at"]